slack helper module
